requirements.txt
streamlit
requests
speechrecognition
pyttsx3
textblob
flask
pinecone-client
sentence-transformers
